Switch to async/await pattern and update code to modern standards
