+++ src/ps4debug/commands.py
@@ -1,0 +1,41 @@
+"""
+PS4Debug raw commands
+"""
+
+VERSION = 0xBD000001
+PROC_LIST = 0xBDAA0001
+PROC_READ = 0xBDAA0002
+PROC_WRITE = 0xBDAA0003
+PROC_MAPS = 0xBDAA0004
+PROC_INSTALL = 0xBDAA0005
+PROC_CALL = 0xBDAA0006
+PROC_ELF = 0xBDAA0007
+PROC_PROTECT = 0xBDAA0008
+PROC_SCAN = 0xBDAA0009
+PROC_INFO = 0xBDAA000A
+PROC_ALLOC = 0xBDAA000B
+PROC_FREE = 0xBDAA000C
+DEBUG_ATTACH = 0xBDBB0001
+DEBUG_DETACH = 0xBDBB0002
+DEBUG_BREAKPT = 0xBDBB0003
+DEBUG_WATCHPT = 0xBDBB0004
+DEBUG_THREADS = 0xBDBB0005
+DEBUG_STOPTHR = 0xBDBB0006
+DEBUG_RESUMETHR = 0xBDBB0007
+DEBUG_GETREGS = 0xBDBB0008
+DEBUG_SETREGS = 0xBDBB0009
+DEBUG_GETFPREGS = 0xBDBB000A
+DEBUG_SETFPREGS = 0xBDBB000B
+DEBUG_GETDBGREGS = 0xBDBB000C
+DEBUG_SETDBGREGS = 0xBDBB000D
+DEBUG_STOPGO = 0xBDBB0010
+DEBUG_THRINFO = 0xBDBB0011
+DEBUG_SINGLESTEP = 0xBDBB0012
+KERN_BASE = 0xBDCC0001
+KERN_READ = 0xBDCC0002
+KERN_WRITE = 0xBDCC0003
+CONSOLE_REBOOT = 0xBDDD0001
+CONSOLE_END = 0xBDDD0002
+CONSOLE_PRINT = 0xBDDD0003
+CONSOLE_NOTIFY = 0xBDDD0004
+CONSOLE_INFO = 0xBDDD0005
