+++ src/ps4debug/exceptions.py
@@ -1,0 +1,3 @@
+class PS4DebugException(Exception):
+    def __init__(self, message):
+        super(PS4DebugException, self).__init__(message)
